requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py36_0
asn1crypto=0.24.0=py36_0
attrs=18.1.0=py36_0
backcall=0.1.0=py_0
blas=1.0=mkl
bleach=2.1.3=py_0
bokeh=0.12.16=py36_0
ca-certificates=2018.4.16=0
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h342bebf_0
chardet=3.0.4=py36h96c241c_1
click=6.7=py36hec950be_0
cloudpickle=0.5.3=py36_0
cryptography=2.2.2=py36h1de35cc_0
cycler=0.10.0=py36_0
cymem=1.31.2=py36_0
cython=0.28.2=py36h1de35cc_0
cytoolz=0.8.2=py36h290905f_0
dask=0.17.5=py36_0
dask-core=0.17.5=py36_0
decorator=4.3.0=py_0
dill=0.2.7.1=py36hdf89ffe_0
distributed=1.21.8=py36_0
entrypoints=0.2.3=py36_1
freetype=2.8=h12048fb_1
ftfy=4.4.3=py36_0
h5py=2.8.0=py36h470a237_0
hdf5=1.10.1=2
heapdict=1.0.0=py36_2
html5lib=1.0.1=py_0
icu=58.2=0
idna=2.6=py36h8628d0a_1
imageio=2.3.0=py36_0
intel-openmp=2018.0.0=8
ipykernel=4.8.2=py36_0
ipython=6.4.0=py36_0
ipython_genutils=0.2.0=py36_0
jedi=0.12.0=py36_0
jinja2=2.10=py36_0
joblib=0.11=py36_0
jpeg=9b=he5867d9_2
jsonschema=2.6.0=py36_1
jupyter_client=5.2.3=py36_0
jupyter_contrib_core=0.3.3=py36_1
jupyter_contrib_nbextensions=0.5.0=py36_0
jupyter_core=4.4.0=py_0
jupyter_highlight_selected_word=0.2.0=py36_0
jupyter_latex_envs=1.4.4=py36_0
jupyter_nbextensions_configurator=0.4.0=py36_0
kiwisolver=1.0.1=py36_1
libcxx=4.0.1=h579ed51_0
libcxxabi=4.0.1=hebd6815_0
libedit=3.1.20170329=hb402a30_2
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libgpuarray=0.7.6=0
libiconv=1.15=0
libpng=1.6.34=he12f830_0
libsodium=1.0.16=0
libtiff=4.0.9=hcb84e12_1
libxml2=2.9.8=0
libxslt=1.1.32=0
locket=0.2.0=py36hca03003_1
lxml=4.2.1=py36_0
mako=1.0.7=py36_0
markupsafe=1.0=py36_0
matplotlib=2.2.2=py36ha7267d0_0
mistune=0.8.3=py36_1
mkl=2018.0.2=1
mkl_fft=1.0.1=py36h917ab60_0
mkl_random=1.0.1=py36h78cc56f_0
more-itertools=4.1.0=py36_0
msgpack-numpy=0.4.1=py36h98f623a_0
msgpack-python=0.4.8=py36h46767b2_0
murmurhash=0.28.0=py36_0
nbconvert=5.3.1=py_1
nbformat=4.4.0=py36_0
ncurses=6.1=h0a44026_0
ninja=1.8.2=py36h04f5b5a_1
nltk=3.3.0=py36_0
notebook=5.5.0=py36_0
numpy=1.14.3=py36h9bb19eb_1
numpy-base=1.14.3=py36h479e554_1
olefile=0.45.1=py36_0
openssl=1.0.2o=0
packaging=17.1=py36_0
pandas=0.23.0=py36_1
pandoc=2.2.1=hde52d81_0
pandocfilters=1.4.2=py36_0
parso=0.2.1=py_0
partd=0.3.8=py36hf5c4cb8_0
patsy=0.5.0=py36_0
pexpect=4.5.0=py36_0
pickleshare=0.7.4=py36_0
pillow=5.1.0=py36hfcce615_0
pip=10.0.1=py36_0
plac=0.9.6=py36_0
pluggy=0.6.0=py36hb1d0581_0
preshed=1.0.0=py36_0
prompt_toolkit=1.0.15=py36_0
psutil=5.4.5=py36h1de35cc_0
ptyprocess=0.5.2=py36_0
py=1.5.3=py36_0
pycparser=2.18=py36h724b2fc_1
pygments=2.2.0=py36_0
pygpu=0.7.6=py36_0
pymc3=3.4.1=py36_0
pyopenssl=18.0.0=py36_0
pyparsing=2.2.0=py36_0
pysocks=1.6.8=py36_0
pytest=3.5.1=py36_0
python=3.6.5=hc167b69_1
python-dateutil=2.7.3=py_0
pytorch=0.4.0=py36_cuda0.0_cudnn0.0_1
pytz=2018.4=py_0
pyyaml=3.12=py36_1
pyzmq=17.0.0=py36_4
readline=7.0=hc1231fa_4
regex=2017.11.09=py36hb672b4b_0
requests=2.18.4=py36h4516966_1
scipy=1.1.0=py36hcaad992_0
send2trash=1.5.0=py_0
setuptools=39.1.0=py36_0
simplegeneric=0.8.1=py36_0
six=1.11.0=py36h0e22d5e_1
sortedcontainers=2.0.3=py36_0
spacy=2.0.11=py36_0
sqlite=3.23.1=hf1716c9_0
tblib=1.3.2=py36hda67792_0
termcolor=1.1.0=py36_1
terminado=0.8.1=py36_0
testpath=0.3.1=py36_0
theano=1.0.1=py36_1
thinc=6.10.2=py36h1702cab_0
tk=8.6.7=h35a86e2_3
toolz=0.9.0=py36_0
torchvision=0.2.1=py36_1
tornado=5.0.2=py36_0
tqdm=4.23.3=py_0
traitlets=4.3.2=py36_0
ujson=1.35=py36_0
urllib3=1.22=py36h68b9469_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_0
wheel=0.31.0=py36_0
wrapt=1.10.11=py36hc29e774_0
xz=5.2.3=h727817e_4
yaml=0.1.7=0
zeromq=4.2.5=1
zict=0.1.3=py36h71da714_0
zlib=1.2.11=hf3cbc9b_2
